# chore: migrate to rspack for build performance