… package.json
… vscode-extension/webpack.config.js