main:

```

> @somesass/project@1.0.0 build
> nx run-many -t build


   ✔  nx run @somesass/vscode-css-languageservice:build (7s)
   ✔  nx run @somesass/language-services:build (1s)
   ✔  nx run some-sass-language-server:build (9s)
   ✔  nx run some-sass:build (3s)

———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————

 NX   Successfully ran target build for 4 projects (20s)
```

rspack:

```
> @somesass/project@1.0.0 build
> nx run-many -t build


   ✔  nx run @somesass/vscode-css-languageservice:build (4s)
   ✔  nx run @somesass/language-services:build (1s)
   ✔  nx run some-sass-language-server:build (1s)
   ✔  nx run some-sass:build (1s)

———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————

 NX   Successfully ran target build for 4 projects (8s)
```

Closes #169